Modified filesystem cache to be service agnostic
